backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
pydantic-settings>=2.2.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
sqlalchemy>=2.0.29
asyncpg>=0.29.0
supabase>=2.4.0
sentence-transformers>=2.7.0
faiss-cpu>=1.8.0
numpy>=1.26.0
pandas>=2.2.0
python-dotenv>=1.0.0
httpx>=0.27.0
openai>=1.0.0
email-validator
rank-bm25>=0.2.2
langsmith
langchain
langchain-openai
openai-whisper>=20231117
PyPDF2>=3.0.1

